Migrate to ThorVG for SVG rendering, add context menu service
- Replaced lunasvg/nanosvg with ThorVG vector graphics engine
- Added right-click context menu service (menu.elf, blob index 6)
- Migrated menubar and dock services from lunasvg to ThorVG
- Fixed premultiplied alpha issue (ARGB8888S non-premultiplied colorspace)
- Fixed R/B byte order swap in menubar and dock raster copy
- Cached SVG raster for fast hover response in menu service
- Simplified Menu.svg and menubar.svg (removed problematic filters/blend modes)
- Removed thread_local usage for freestanding environment compatibility
- Persistent canvas per ThorVG doc to avoid mpool stale pointer crashes
- Embedded Arial TTF font for SVG text rendering
- Cleaned up lunasvg/nanosvg directories and Makefile rules
- Moved cxx_runtime.cpp to shared userspace/lib/cpp_runtime/
- Updated README with new screenshot and project layout

+++ kernel/include/syscall.h
@@ -240,7 +240,7 @@ int sys_fstat(int fd, void *statbuf);
 int sys_unlink(const char *path);
 int sys_getcwd(char *buf, size_t size);
 int sys_chdir(const char *path);
-int sys_brk(void *addr);
+uint64_t sys_brk(void *addr);
 
 /* POSIX socket wrappers */
 int sys_socket(int domain, int type, int protocol);
